Consider the padding in the input width calculation
The offset parent element may have a padding set which leads to a input
size result that is too big. Therefore the input element is always wrapped
to a new line instead using the remaining space.

Considering the padding is not easy: In a general case we have to read
the calculated css styles and parse the string containing units. But here
we use the intermediate div element and shortcut using its bounding box
that already have the padding of its parent element subtracted.